Autolinking of speakers in sessions page ready.
If "John Hacker" is the speaker that has to be autolinked in the
sessions MD file; the markup to autolink is [John Hacker]().

[John Hacker]() -> <a href="speakers.html#hacker">John Hacker</a>, if id
for "John Hacker" is available in the `speaker.ids` file[1].

[1]: The `speakers.ids` file is automatically written to the disk (in
the current working directory) when the speakers' bio page is generated.

List of speakers that are autolinked but don't have an id, are written
to the `speakers.noid` file when the sessions page is generated.

Addresses issue #7.